% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metIdentify.R
\name{filterIden}
\alias{filterIden}
\title{filterIden}
\usage{
filterIden(object, ms1.match.ppm = 25, rt.match.tol = 30,
  ms2.match.tol = 0.5, total.score.tol = 0.5)
}
\arguments{
\item{object}{A metIdentifyClass object.}

\item{ms1.match.ppm}{MS1 match ppm.}

\item{rt.match.tol}{RT match tolerance.}

\item{ms2.match.tol}{MS2 match tolerance.}

\item{total.score.tol}{Total score tolerance.}
}
\value{
An new metIdentifyClass.
}
\description{
Filter identifications according to mz error, RT error, MS similarity and total score.
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
