% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metIdentify.R
\name{getIdenInfo}
\alias{getIdenInfo}
\title{getIdenInfo}
\usage{
getIdenInfo(object, which.peak, database)
}
\arguments{
\item{object}{A metIdentifyClass object.}

\item{which.peak}{Peak name or "all".}

\item{database}{Database used.}
}
\value{
A identification table (data.frame).
}
\description{
Get identification information from a metIdentifyClass object.
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
