% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metIdentify.R
\name{getIdentificationTable}
\alias{getIdentificationTable}
\title{getIdentificationTable}
\usage{
getIdentificationTable(object, type = c("old", "new"))
}
\arguments{
\item{object}{A metIdentifyClass object.}

\item{type}{The stype of identification table.}
}
\value{
A identification table (data.frame).
}
\description{
Get identification table from a metIdentifyClass object.
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
