% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databaseConstruction.R
\name{getMS2spectrum}
\alias{getMS2spectrum}
\title{getMS2spectrum}
\usage{
getMS2spectrum(lab.id, database, polarity = c("positive", "negative"),
  ce = "30")
}
\arguments{
\item{lab.id}{Lab ID.}

\item{database}{Database.}

\item{polarity}{positive or negative.}

\item{ce}{Collision value.}
}
\value{
A metIdentifyClass object.
}
\description{
Get Ms2 spectra of peaks from metIdentifyClass.
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
