% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function4metIdentifyClass.R
\name{getMS2spectrum2Object}
\alias{getMS2spectrum2Object}
\title{getMS2spectrum2Object}
\usage{
getMS2spectrum2Object(peak.name, object)
}
\arguments{
\item{peak.name}{Peak name.}

\item{object}{metIdentifyClass}
}
\value{
A MS2 spectrum.
}
\description{
Get spectra of peaks.
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
