% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzIdentify.R
\name{getParams2}
\alias{getParams2}
\title{getParams2}
\usage{
getParams2(object)
}
\arguments{
\item{object}{A mzIdentifyClass object.}
}
\value{
A data.frame contains all the parameters of this metIdentifiyClass object.
}
\description{
Get parameters from a mzIdentifyClass object.
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
