% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metIdentify.R
\name{metIdentify}
\alias{metIdentify}
\title{metIdentify}
\usage{
metIdentify(ms1.data, ms2.data, ms1.ms2.match.mz.tol = 25,
  ms1.ms2.match.rt.tol = 10, ms1.match.ppm = 25, ms2.match.ppm = 30,
  mz.ppm.thr = 400, ms2.match.tol = 0.5, fraction.weight = 0.3,
  dp.forward.weight = 0.6, dp.reverse.weight = 0.1,
  rt.match.tol = 30, polarity = c("positive", "negative"), ce = "30",
  column = c("hilic", "rp"), ms1.match.weight = 0.25,
  rt.match.weight = 0.25, ms2.match.weight = 0.5, path = ".",
  total.score.tol = 0.5, candidate.num = 3, database, threads = 3)
}
\arguments{
\item{ms1.data}{The name of ms1 peak table (csv format). Column 1 is "name", Column 2 is
"mz" and column is "rt" (second).}

\item{ms2.data}{MS2 data, they must be mgf of msp files. For example, ms2.data = c("test.mgf", "test2.msp").}

\item{ms1.ms2.match.mz.tol}{MS1 peak and MS2 spectrum matching m/z tolerance.}

\item{ms1.ms2.match.rt.tol}{MS1 peak and MS2 spectrum matching RT tolerance.}

\item{ms1.match.ppm}{Precursor match ppm tolerance.}

\item{ms2.match.ppm}{Fragment ion match ppm tolerance.}

\item{mz.ppm.thr}{m/z}

\item{ms2.match.tol}{MS2 match tolerance.}

\item{fraction.weight}{Fraction weight.}

\item{dp.forward.weight}{Forward DP weight.}

\item{dp.reverse.weight}{Reverse DP weight.}

\item{rt.match.tol}{RT match tolerance.}

\item{polarity}{The polarity of data, "positive"or "negative".}

\item{ce}{Collision energy.}

\item{column}{"hilic" or "rp".}

\item{ms1.match.weight}{MS1 match weight.}

\item{rt.match.weight}{RT match weight.}

\item{ms2.match.weight}{MS2 match weight.}

\item{path}{Work directory.}

\item{total.score.tol}{Total score tolerance.}

\item{candidate.num}{The number of candidate.}

\item{database}{MS2 database name.}

\item{threads}{Number of threads}
}
\value{
A metIdentifyClass object.
}
\description{
Identify peaks based on MS/MS database.
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
