% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function4metIdentifyClass.R
\name{whichHasIden}
\alias{whichHasIden}
\title{whichHasIden}
\usage{
whichHasIden(object)
}
\arguments{
\item{object}{A metIdentifyClass object.}
}
\value{
Spectra names with identifications.
}
\description{
Get the spectra names with identifications.
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
