% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIdentificationTable.R
\name{getIdentificationTable}
\alias{getIdentificationTable}
\title{getIdentificationTable}
\usage{
getIdentificationTable(..., candidate.num = 3, type = c("old", "new"))
}
\arguments{
\item{...}{One or multiple metIdentifyClass objects.}

\item{candidate.num}{The number of camdidates.}

\item{type}{The stype of identification table.}
}
\value{
A identification table (data.frame).
}
\description{
Get identification table from a metIdentifyClass object.
}
\author{
Xiaotao Shen
\email{shenxt1990@163.com}
}
